+++ examples/hall/route/user.go
@@ -7,8 +7,8 @@ import (
 
 type User struct {
 	userID        uint32
-	gateID        uint32
-	hallID        uint32
+	gateID        uint8
+	hallID        uint8
 	roomID        uint32
 	*network.Conn // 玩家可能从不同的网关过来，所以需要存一下网关ID
 }
@@ -17,18 +17,18 @@ func (u *User) UserID() uint32 {
 	return u.userID
 }
 
-func (u *User) GameID() uint32 {
+func (u *User) GameID() uint8 {
 	return 0
 }
 
-func (u *User) GateID() uint32 {
+func (u *User) GateID() uint8 {
 	return u.gateID
 }
 
 type RoomTemplete struct {
 	TempId    uint32
 	UserCount uint32
-	GameID    uint32
+	GameID    uint8
 }
 
 type RoomInstance struct {
